# perf: GetObjects call is slow even when filters are provided