## Replaced some cursor calls with static calls and filtered early when possible

If a share isn't associated with a database then SELECT DATABASE_NAME FROM INFORMATION_SCHEMA.DATABASES call will not list it and therefore it doesn't seem necessary to call SHOW SHARES LIKE '%database_name%' to get a list and check if a DB isn't created for it. Therefore, those checks were removed.

### Comparison of performance improvements:

I have created another PR: https://github.com/apache/arrow-adbc/pull/1299 with additional tests for `GetObjects`.

| Test | Before | After | % Improvement |
|-----|---------------------|------------------|-------------------|
| CanGetObjectsAll | 17.6 s | 3 s | 82.5  |
| CanGetObjectsCatalogs | 503 ms | 333 ms | 33.80 | 
| CanGetObjectsCatalogsWithPattern | 421 ms | 369 ms | 12.35 | 
| CanGetObjectsDbSchemas | 4.4 s | 694 ms | 84.36 |
| CanGetObjectsDbSchemasWithPattern | 4 s | 807 ms | 79.825 |
| CanGetObjectsTables | 18 s | 2.8 s | 84.44 | 
| CanGetObjectsTablesWithPattern | 17.6 s | 2.9 s | 83.52 |