@jduo 2023-11-15
It'd be great to see what performance numbers are looking like after these changes.