@ryan-syed 2023-11-16
> It'd be great to see what performance numbers are looking like after these changes.

Added the performance numbers in the PR description.