@lidavidm 2023-11-20
@zeroshade in the future can we make sure this is `perf(go/adbc/driver/snowflake): ...` or similar